# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для записи к специалистам",
        "title": "LAPS API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/specialists": {
            "get": {
                "description": "Возвращает список специалистов с фильтрацией и пагинацией",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Специалисты"
                ],
                "summary": "Получить список специалистов",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Лимит записей на странице (по умолчанию 20)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Смещение (по умолчанию 0)",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Тип специалиста (lawyer, psychologist, nutritionist, trainer)",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "ID специализации",
                        "name": "specialization_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Дата для получения свободных слотов (YYYY-MM-DD)",
                        "name": "date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список специалистов с пагинацией",
                        "schema": {
                            "$ref": "#/definitions/rest.paginatedResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Создает или обновляет профиль специалиста для пользователя (при регистрации профиль создается автоматически)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Специалисты"
                ],
                "summary": "Создать специалиста",
                "parameters": [
                    {
                        "description": "Данные специалиста",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CreateSpecialistDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "ID созданного специалиста",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    }
                }
            }
        },
        "/specialists/me": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает профиль специалиста для текущего авторизованного пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Специалисты"
                ],
                "summary": "Получить профиль специалиста текущего пользователя",
                "responses": {
                    "200": {
                        "description": "Данные специалиста",
                        "schema": {
                            "$ref": "#/definitions/domain.Specialist"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "404": {
                        "description": "Профиль специалиста не найден",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    }
                }
            }
        },
        "/specialists/{id}": {
            "get": {
                "description": "Возвращает информацию о специалисте по указанному ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Специалисты"
                ],
                "summary": "Получить специалиста по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID специалиста",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные специалиста",
                        "schema": {
                            "$ref": "#/definitions/domain.Specialist"
                        }
                    },
                    "400": {
                        "description": "Неверный формат ID",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "404": {
                        "description": "Специалист не найден",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Обновляет информацию о специалисте",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Специалисты"
                ],
                "summary": "Обновить специалиста",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID специалиста",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новые данные специалиста",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.UpdateSpecialistDTO"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Данные успешно обновлены"
                    },
                    "400": {
                        "description": "Ошибка валидации",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "404": {
                        "description": "Специалист не найден",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Удаляет профиль специалиста",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Специалисты"
                ],
                "summary": "Удалить специалиста",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID специалиста",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Профиль специалиста удален"
                    },
                    "400": {
                        "description": "Неверный формат ID",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "404": {
                        "description": "Специалист не найден",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    }
                }
            }
        },
...
        "/specialists/{id}/reviews": {
            "get": {
                "description": "Возвращает список отзывов о специалисте с пагинацией (перенаправляет на /reviews)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Специалисты",
                    "Отзывы"
                ],
                "summary": "Получить отзывы о специалисте",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID специалиста",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Лимит записей на странице (по умолчанию 10)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Смещение (по умолчанию 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список отзывов с пагинацией",
                        "schema": {
                            "$ref": "#/definitions/rest.paginatedResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный формат ID",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/rest.errorResponseBody"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.CreateSpecialistDTO": {
            "type": "object",
            "properties": {
                "association_member": {
                    "type": "boolean"
                },
                "description": {
                    "type": "string"
                },
                "education": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.EducationDTO"
                    }
                },
                "experience": {
                    "type": "integer",
                    "minimum": 0
                },
                "experience_years": {
                    "type": "integer"
                },
                "primary_consult_price": {
                    "type": "number",
                    "minimum": 0
                },
                "secondary_consult_price": {
                    "type": "number",
                    "minimum": 0
                },
                "specialization_id": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                },
                "work_experience": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.WorkExperienceDTO"
                    }
                }
            }
        },
...
        "domain.EducationDTO": {
            "type": "object",
            "required": [
                "degree",
                "graduation_year",
                "institution",
                "specialization"
            ],
            "properties": {
                "degree": {
                    "type": "string"
                },
                "graduation_year": {
                    "type": "integer"
                },
                "institution": {
                    "type": "string"
                },
                "specialization": {
                    "type": "string"
                }
            }
        },
...
        "rest.errorResponseBody": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "rest.paginatedResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total_count": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}